Refactor to use vfsutil.WalkFiles to avoid accessing files twice.
This is an internal generation code change, the generated output should
be computed more efficiently but the output should be unchanged.

When using vfsutil.Walk, each file in the input virtual filesystem was
opened twice. First time just to perform a Stat, and second time when
actually copying its contents.

Switch to using vfsutil.WalkFiles, which allows walking over files and
provides file content as a ReadSeeker.

This is helpful when walking virtual filesystems that perform
computation on each Open call (even to do a Stat, since the file size
needs to be calculated).

Don't use bufio.NewWriter for now because I will want to potentially do
some seeking and truncating (for not-worth-compressing files). It can
be re-added later if needed.