Avoid compressing files that do not compress well.
By default, vfsgen writes out all assets with gzip compressed content.
When doing so, it's possible to detect when gzip compression is
detrimental and does not make the file smaller.

Detect that, rewind and rewrite the asset without compression (it will
be smaller this way).

This has the advantage of making the generated Go code smaller in size
and faster at runtime. Generating may take a little longer as some
files may need to be written twice.

This has another advantage, the information that certain files are not
worth gzip compressing can be encoded. Then, e.g., when serving assets,
it can be determined that this file wouldn't become smaller when gzip
compressed, so that can be avoided without additional computational
cost.

Add _vfsgen_fileInfo and _vfsgen_file for representing files that are
not worth gzip compressing.

Update and add tests.