README: Add Alternatives and Comparison sections.
This is good for users because they can either be more confident that
vfsgen is the best choice for them, or if their needs are different,
find a more well-suited package.

It's good for vfsgen because it communicates to users more clearly that
it is indeed similar to some other existing packages.

It's good for alternative packages listed because they get free
advertisement and exposure.

It's good for everyone because it can help reduce duplication of effort
if something can be consolidated, and improved code quality and
functionality due to competition. :)