Make the VariableName option default to being exported.
This change makes the VariableName default to the exported `Assets` instead of
the previously unexported `assets`. The rational for this is that people creating
Go packages will not have to specify a VariableName parameter at all because it
will default to being exported. For creators of `main` packages, they cannot be
imported anyway so keeping the field private by default is less valuable.