Add ./cmd/vfsgendev binary to address common use case.
Add an opinionated binary for specialized use. It uses the convention
of using "dev" build tag for development mode, and assumes that the
source VFS is defined in a "dev" only environment; it uses "!dev" build
tag in output .go file.

Usage is simple:

	vfsgendev -source="import/path".VariableName

See sourcegraph/appdash#91 for additional background, motivation, and
discussion.

Resolves #2.

Closes #14.