Generate separate data and common files.
The goal is to resolve #23, to make it possible to use vfsgen multiple
times in same package.

Previously, that would cause an error because the same types would be
defined in each generated file.

The approach to solve that is to generate two files:

-	VFS data file
-	common type definitions file

This avoids the multiple declaration issues in an efficient way.

A downside is having 2 files as generated output, instead of one.