Simplify encoding of time values.
We can directly generate a time.Date constructor rather than
going through the text marshaling/unmarshaling mechanism.

This is simpler. The reason it was being done via Time.MarshalText
and mustUnmarshalTextTime before is because I hadn't thought of this.

It also happens to fix a "mustUnmarshalTextTime declared and not used"
error when the generated filesystem is empty (issue #42).

Fixes #42.
Updates a2edb6035b1469bc3aac9ca58fbef9b61be9876e.