Generate output in-memory, then write all at once.
Users should have plenty of memory to hold the entire output as its
being generated. There's no need to start writing it to disk while
generation is still in progress.

So, generate the entire output in-memory, then write it all at once to
disk.

This should be a better user experience because it prevents the user
from potentially seeing or being disturbed by partial output while the
generation process is running.

Resolves #35.