Update generated code for Go 1.11 gofmt change.
Go 1.11 has been released today. The gofmt formatting algorithm has
undergone minor tweaks (see https://golang.org/doc/go1.11#gofmt),
which happen to affect this particular struct literal instance.

Update it so that it's compatible with gofmt of Go 1.11, the
current release. It's not going to be compatible with Go 1.10 or
older, but that is in line with the support policy described at
https://github.com/shurcooL/SLA#go-version.

In general, as mentioned at https://golang.org/doc/go1.11#gofmt,
minor updates to gofmt are expected from time to time, and any
continuous integration checks for gofmt compatibility need to
pick a single Go version that they expect conformity with. It
may not always be viable to be gofmt compatible with multiple
Go versions at the same time in the general case. This time,
it was possible, but only by compromising code readability,
which I decided against doing to avoid setting poor precedent.

As a result, users of this version of vfsgen need to expect
the generated code to be compatible with Go 1.11 only.
Otherwise, they can run a gofmt version of their choice
as a follow-up step after generating code with vfsgen.

Reference: https://golang.org/doc/go1.11#gofmt.