vfsgendev: module aware source lookup
Currently vfsgendev only considers GOROOT and GOPATH when looking up the
source to use. This does not work as expected when running vfsgendev
within a go module, as introduced in Go 1.11.

Specifying the absolute path of the current directory instead of
the empty path ("") as the source directory for build.Import() allows
the module import logic to work when running inside a go module.

As a non-empty a source directory would allow passing relative import
paths to the -source option, this is checked for and explicitly
forbidden.