Error if input filesystem is not readable. (#48)
If the input filesystem is not readable, previously we would log that
to stderr and keep going. This behavior certainly doesn't make sense
when the error happens to the root directory of the input filesystem,
so fix that and add a test case for it.

However, it's probably also good to be strict for other files inside
the input directory. Do that for now. If it causes legitimate issues,
we can relax it to be less strict (i.e., only treat error reading root
directory as fatal, but not the rest).