remove unconditional write to stdout
Writiting to the stdout unconditionally is noisy and inflexible.
The caller is in a better position to decide if anything should
be printed to stdout, and it can already do that. So remove the
uncoditional write from the Generate function.

Don't add anything to vfsgendev, keeping it quiet when everything
is successful. This is consistent with various Go code generators
like stringer, and testing it out over the last few weeks did not
make me feel like something is missing. It's also more consistent
with the philosophy for Unix and Plan 9 tools, where it's nice to
be quiet when all goes as planned.

For https://github.com/shurcooL/vfsgen/issues/81.